requirements.txt
botbuilder-integration-aiohttp>=4.9.1
pynetbox